docs: rename fields.py to _fields.py so `bagof.magic.fields` is the function
Maintainer decision on the question raised in the previous commit here
(see issue #6): rename the submodule rather than the top-level
`fields()` function.

Griffe's static package loader always re-registers a submodule onto
its parent when it loads that submodule's file, unconditionally
overwriting whatever member of the same name was already there --
including a function defined earlier in `__init__.py`. That made
`fields()` permanently unreachable at the dotted path
`bagof.magic.fields` in the API reference, no matter what mkdocstrings
page options were passed (griffe's own loader even logs a "submodule
is shadowing the member at the same path" warning for exactly this
case).

Renaming `fields.py` to `_fields.py` frees the name: nothing in
`bagof.magic`'s public surface changes (`from .fields import *` /
`Field` / `__all__ as __all_fields__` become the same imports from
`._fields`; every name a user imports -- `Field`, `Factory`, `Doc`,
etc. -- still resolves the same way, and `Factory`'s own `__all__` on
the private module still makes its members public in griffe's model,
so `docs/api/fields.md` needed no filters override, only pointing
`::: bagof.magic.fields` at `::: bagof.magic._fields`), but
`bagof.magic.fields` is no longer claimed by a submodule -- confirmed
with `python -c` that `bagof.magic.fields` is the function at runtime
(the shadowing submodule import now genuinely fails: `import
bagof.magic.fields` raises `ModuleNotFoundError`, as intended).

`fields` is now in the explicit `members:` list on `docs/api/index.md`
added in the previous commit, and a real `zensical build -c` confirms
it renders there with its signature, docstring, and a working
cross-reference to `Field` -- the actual point of the whole rename.

Also updated the two other places that named `fields.py` by its old
name: `tests/test_docstrings.py`'s `SOURCES` list (the file that scans
docstrings for `pycon` blocks -- otherwise its examples silently stop
being checked) and CLAUDE.md's layout/doc-style sections. Confirmed
`tests/test_docstrings.py` still collects 28 `pycon`-block tests (well
above its own "found at least twenty" sanity check) after the rename.

Verified: full test suite (287 passed), ruff clean (after re-sorting
the now `_`-prefixed import ahead of `.constants` alphabetically), and
a real `zensical build -c` showing both `MetaMagic` and `fields` on
the API reference's front page.

Fixes #2
Fixes #6
