+++ docs/api/fields.md
@@ -1,1 +1,1 @@
-# ::: bagof.magic.fields
+# ::: bagof.magic._fields
