+++ src/bagof/magic/_fields.py
