+++ tests/test_magic.py
@@ -1132,7 +1132,7 @@ class TestUtils:
 
 
 # ======================================================================
-# fields.py: Field internals
+# _fields.py: Field internals
 # ======================================================================
